# api/api.yaml
openapi: 3.0.1
info:
  title: Generic Dmaap and Kafka Information Producer
  description: Reads data from DMaaP and Kafka and posts it further to information
    consumers
  license:
    name: Copyright (C) 2021 Nordix Foundation. Licensed under the Apache License.
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
servers:
- url: /
...
paths:
...
  /generic_dataproducer/info_job:
    get:
      tags:
      - Producer job control API
      summary: Get all jobs
      description: Returns all info jobs, can be used for trouble shooting
      operationId: getJobs
      responses:
        200:
          description: Information jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/producer_info_job_request'
    post:
      tags:
      - Producer job control API
      summary: Callback for Information Job creation/modification
      description: The call is invoked to activate or to modify a data subscription.
        The endpoint is provided by the Information Producer.
      operationId: jobCreatedCallback
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/void'
        400:
          description: Other error in the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_information'
        404:
          description: Information type is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_information'
...
  /generic_dataproducer/info_job/{infoJobId}:
    delete:
      tags:
      - Producer job control API
      summary: Callback for Information Job deletion
      description: The call is invoked to terminate a data subscription. The endpoint
        is provided by the Information Producer.
      operationId: jobDeletedCallback
      parameters:
      - name: infoJobId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/void'
...
components:
  schemas:
    producer_info_job_request:
      required:
      - info_job_identity
      type: object
      properties:
        owner:
          type: string
          description: The owner of the job
        last_updated:
          type: string
          description: The time when the job was last updated or created (ISO-8601)
        info_job_identity:
          type: string
          description: Identity of the Information Job
        target_uri:
          type: string
          description: URI for the target of the produced Information
        info_job_data:
          type: object
          description: Json for the job data
        info_type_identity:
          type: string
          description: Type identity for the job
      description: The body of the Information Producer callbacks for Information
        Job creation and deletion
    error_information:
      type: object
      properties:
        detail:
          type: string
          description: ' A human-readable explanation specific to this occurrence
            of the problem.'
          example: Policy type not found
        status:
          type: integer
          description: 'The HTTP status code generated by the origin server for this
            occurrence of the problem. '
          format: int32
          example: 503
      description: Problem as defined in https://tools.ietf.org/html/rfc7807
    void:
      type: object
      description: Void/empty
...